qed.nexus: the recipe, identity, and worker-side reconstruction core moves from Tile to Chore
